requirements.txt
pandas
openpyxl
tqdm
pymupdf
pytesseract
Pillow
streamlit
watchdog
